move drawing functions from map component
